move: scheduler 한 곳으로 이동
